optimize API related to item movement

+++ game_control/game_interface/item_renaming.go
@@ -143,7 +143,7 @@ func (g *GameInterface) RenameItem(
 		return &AnvilOperationResponse{
 			Successful: false,
 			Destination: &ItemLocation{
-				WindowID:    int16(containerOpeningData.WindowID),
+				WindowID:    containerOpeningData.WindowID,
 				ContainerID: 0,
 				Slot:        1,
 			},
